requirements.txt
faiss-gpu==1.7.2
tqdm==4.67.1
pillow==11.0.0
scikit-learn==1.6.1
numpy==1.23.0
pytorch-metric-learning==2.8.1
pandas==2.2.3
einops==0.8.1
